==> 3 ludonauta to airtable.ipynb <==
import numpy as np
import pandas as pd
import requests
from pprint import pprint

# Es necesario agregar tu propio token si quieres ver el funcionamiento de subida de datos a airtable

base_id = "appyddcaGTw6if8vV"
token_id = "tu token"
table_id = "tbl5DkDouUihJEShO"

airtable_url = "https://api.airtable.com/v0/"


endpoint = f"https://api.airtable.com/v0/{base_id}/{table_id}"

endpoint

headers = {
    "Authorization":f"Bearer {token_id} ",
    "Content-Type": "application/json"
}

pprint(headers)

def convertir_a_lista(cadena):
    return eval(cadena)

df = pd.read_csv('juegos_ludonauta_limpios.csv', converters={'Autores': convertir_a_lista,
                                                                    'Categorias': convertir_a_lista,
                                                                    'Mecanicas' : convertir_a_lista,
                                                                    'Juegos relacionados' : convertir_a_lista})

df.isna().sum()

df[df["Link"].isna()]

response = requests.get(url = endpoint, headers = headers)

response.json()

pprint({"records" : [{"fields" : {"Link" : "ejmeplo",
                                  "Nombre" : "nombre_juego",
                                  "Fecha Publicación" : 2022}}]})

datos_subir = {"records" : [{"fields" : df.iloc[0, :].to_dict()}],
               "typecast" : True}

pprint(datos_subir)

response = requests.post(url = endpoint, json = datos_subir, headers = headers)

response.json()

%%time

for i in range(df.shape[0]):

    datos_subir = {"records" : [{"fields" : df.iloc[i, :].to_dict()}],
                   "typecast" : True}
    
    response = requests.post(url = endpoint, json = datos_subir, headers = headers)


base_id = "app7rphQpjYdtFAMM"
token_id = "patYj0ekod7IkqHMq.b59d509464af0746238f3a63625a2e4ee737586f631c218828fffe138c782189"
table_id = "tblyAG4wrWvNAhbOp"

airtable_url = "https://api.airtable.com/v0/"


endpoint = f"https://api.airtable.com/v0/{base_id}/{table_id}"

endpoint

headers = {
    "Authorization":f"Bearer {token_id} ",
    "Content-Type": "application/json"
}

pprint(headers)

df2 = pd.read_csv('precios_limpios.csv', converters={'Nombre tienda': convertir_a_lista,
                                                    'Disponibilidad' : convertir_a_lista})

df2 = df2.drop(['Unnamed: 0', 'Unnamed: 0.1'], axis=1)

df2.isna().sum()

df2[df2['Link'].isna()]

response = requests.get(url = endpoint, headers = headers)

response.json()

pprint({"records" : [{"fields" : {"Link" : "ejmeplo",
                                  "Tienda juego" : "Mundo juegos",
                                  "Precio medio" : 13.45}}]})

datos_subir = {"records" : [{"fields" : df2.iloc[0, :].to_dict()}],
               "typecast" : True}

pprint(datos_subir)

response = requests.post(url = endpoint, json = datos_subir, headers = headers)

response.json()

%%time

for i in range(df2.shape[0]):

    datos_subir = {"records" : [{"fields" : df2.iloc[i, :].to_dict()}],
                   "typecast" : True}
    
    response = requests.post(url = endpoint, json = datos_subir, headers = headers)
