==> 4 ludonauta from Airtable.ipynb <==
import numpy as np
import pandas as pd

import requests
from time import sleep

from pprint import pprint

# Claves para los datos sobre JUEGOS de Ludonauta
# Es necesario agregar tu propio token si quieres ver el funcionamiento de la extracción de datos a airtable

base_id = "appyddcaGTw6if8vV"
token_id = "tu token"
table_id = "tbl5DkDouUihJEShO"

airtable_url = "https://api.airtable.com/v0/"


endpoint = f"https://api.airtable.com/v0/{base_id}/{table_id}"

endpoint

headers = {
    "Authorization":f"Bearer {token_id} ",
    "Content-Type": "application/json"
}

pprint(headers)

response = requests.get(url = endpoint, headers = headers)

response.json()

params = {"offset" : None}

df_airtable = pd.DataFrame()

while params.get("offset") != None or df_airtable.shape[0] == 0:
    
    response = requests.get(url = endpoint, headers = headers, params = params)
    
    print(response.url)
    
    print(f"response: {response.status_code}")
    
    params["offset"] = response.json().get("offset")
    
    print(params.get("offset"))
    
    df_airtable = pd.concat([df_airtable, pd.json_normalize(response.json()["records"])], ignore_index = True)
    
    sleep(0.5)
    
df_airtable.shape[0]

df_juegos = df_airtable

df_juegos = df_juegos.drop(['id', 'createdTime'], axis =1)

df_juegos = df_juegos.rename(columns={'fields.Juegos relacionados' : 'Juegos relacionados',
                                      'fields.Autores' :'Autores',
                                      'fields.Dep idioma' : 'Dep idioma',
                                      'fields.Edad' : 'Edad',
                                      'fields.Jugadores maximos' : 'Jugadores maximos',	
                                      'fields.Duracion maxima' : 'Duracion maxima',	
                                      'fields.Jugadores minimos' : 'Jugadores minimos',
                                      'fields.Nombre' : 'Nombre',	
                                      'fields.Edicion juegos' : 'Edicion juegos', 
                                      'fields.Duracion' : 'Duracion',
                                      'fields.Complejidad' : 'Complejidad',	
                                      'fields.Mediana duracion' : 'Mediana duracion',
                                      'fields.Puntuacion' : 'Puntuacion',	
                                      'fields.Link' : 'Link',	
                                      'fields.Duracion minima' : 'Duracion minima',
                                      'fields.Categorias' :'Categorias',	
                                      'fields.Jugadores': 'Jugadores',	
                                      'fields.Mecanicas' : 'Mecanicas',	
                                      'fields.Fecha publicacion' : 'Fecha publicacion',	
                                      'fields.Mediana jugadores' : 'Mediana jugadores'})

df_juegos.head(1)

df_juegos.to_csv('juegos_ludonauta_airtable.csv')

# Claves para los datos sobre PRECIOS de Ludonauta
base_id = "app7rphQpjYdtFAMM"
token_id = "patYj0ekod7IkqHMq.b59d509464af0746238f3a63625a2e4ee737586f631c218828fffe138c782189"
table_id = "tblyAG4wrWvNAhbOp"

airtable_url = "https://api.airtable.com/v0/"


endpoint = f"https://api.airtable.com/v0/{base_id}/{table_id}"

endpoint

headers = {
    "Authorization":f"Bearer {token_id} ",
    "Content-Type": "application/json"
}

pprint(headers)

response = requests.get(url = endpoint, headers = headers)

response.json()

params = {"offset" : None}

df_airtable = pd.DataFrame()

while params.get("offset") != None or df_airtable.shape[0] == 0:
    
    response = requests.get(url = endpoint, headers = headers, params = params)
    
    print(response.url)
    
    print(f"response: {response.status_code}")
    
    params["offset"] = response.json().get("offset")
    
    print(params.get("offset"))
    
    df_airtable = pd.concat([df_airtable, pd.json_normalize(response.json()["records"])], ignore_index = True)
    
    sleep(0.5)
    
df_airtable.shape[0]

df_precios = df_airtable

df_precios = df_precios.drop(['id', 'createdTime'], axis=1)

df_precios = df_precios.rename(columns={'fields.Nombre tienda': 'Nombre tienda',
                                        'fields.Precio medio' : 'Precio medio',
                                        'fields.Link' : 'Link',
                                        'fields.Precio' : 'Precio',
                                        'fields.Disponibilidad' : 'Disponibilidad'})

df_precios.head()

df_precios.to_csv('precios_ludonauta_airtable.csv')
